// superset-frontend/plugins/plugin-chart-remita-table/src/types.ts
export type VisibilityCondition = 'all' | 'selected' | 'unselected';
export type ActionStyle = 'default' | 'primary' | 'danger';

export interface BulkAction {
  key: string;
  label: string;
  type: 'dropdown' | 'button';
  style?: ActionStyle;
  icon?:
    | 'plus'
    | 'edit'
    | 'delete'
    | 'eye'
    | 'link'
    | 'check'
    | 'key'
    | 'tag'
    | 'more';
  tooltip?: string;
  boundToSelection: boolean;
  visibilityCondition: VisibilityCondition;
  showInSliceHeader: boolean;
  value: any;
  rowId: any
}